Move current history provider, strip nested directories
