--- client/src/components/providers/History/UserHistories/index.js
@@ -1,2 +1,0 @@
-import UserHistories from "./UserHistories";
-export default UserHistories;
