--- client/src/components/providers/History/index.js
@@ -1,7 +1,0 @@
-/**
- * Providers are renderless components that take raw inputs and monitor the
- * cache or other observables to emit end results as slot props.
- */
-
-// Management for current user's histories. Largely a passthrough of store methods
-export { default as UserHistories } from "./UserHistories";
